+++ src/Ui/Bundle/EventBundle/Controller/Happening/BeginReplyHappeningQuestionAction.php
@@ -1,0 +1,87 @@
+<?php
+
+namespace Proximum\Vimeet\Ui\Bundle\EventBundle\Controller\Happening;
+
+use Proximum\Vimeet\Application\Adapter\AuthorizationCheckerAdapterInterface;
+use Proximum\Vimeet\Application\Adapter\CommandBusInterface;
+use Proximum\Vimeet\Application\Command\Happening\Webinar\Question\BeginReplyHappeningQuestion;
+use Proximum\Vimeet\Application\Exception\Happening\HappeningException;
+use Proximum\Vimeet\Application\Exception\Happening\QuestionNotFoundException;
+use Proximum\Vimeet\Application\Exception\Happening\ReplyQuestionNotAllowedException;
+use Proximum\Vimeet\Application\Query\Happening\Webinar\CanAccessToWebinar;
+use Proximum\Vimeet\Domain\Model\Happening;
+use Proximum\Vimeet\Domain\Model\Sheet;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\ParamConverter\EventDomain;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\Security\Happening\ParticipationVoter;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\Security\SheetVoter;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\ValueResolver\UserDomain;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+
+class BeginReplyHappeningQuestionAction
+{
+    /** @var AuthorizationCheckerAdapterInterface */
+    private $authorizationCheckerAdapter;
+
+    /** @var CanAccessToWebinar */
+    private $canAccessToWebinar;
+
+    /** @var CommandBusInterface */
+    private $commandBus;
+
+    public function __construct(
+        AuthorizationCheckerAdapterInterface $authorizationCheckerAdapter,
+        CanAccessToWebinar $canAccessToWebinar,
+        CommandBusInterface $commandBus
+    ) {
+        $this->authorizationCheckerAdapter = $authorizationCheckerAdapter;
+        $this->canAccessToWebinar = $canAccessToWebinar;
+        $this->commandBus = $commandBus;
+    }
+
+    public function __invoke(
+        Request $request,
+        EventDomain $eventDomain,
+        Sheet $sheet,
+        Happening $happening,
+        UserDomain $userDomain
+    ): JsonResponse {
+        $event = $eventDomain->getEvent();
+        $user = $userDomain->getUser();
+
+        if (!$this->authorizationCheckerAdapter->isGranted('IS_AUTHENTICATED_REMEMBERED')
+            || !$this->authorizationCheckerAdapter->isGranted('PERMISSION_HAPPENING_ACCESS', $event)
+            || !$this->authorizationCheckerAdapter->isGranted(SheetVoter::EDIT, $sheet)
+            || !$this->authorizationCheckerAdapter->isGranted(ParticipationVoter::PARTICIPATE, $sheet)
+            || !$this->canAccessToWebinar->isSatisfiableBy($happening, $user)
+            || $happening->getEvent() !== $event
+            || $sheet->getEvent() !== $event
+        ) {
+            throw new AccessDeniedException('Access denied to this happening');
+        }
+
+        $payload = json_decode($request->getContent(), true);
+        $questionId = (int) ($payload['questionId'] ?? null);
+
+        if (!$questionId) {
+            return new JsonResponse(['status' => 'error', 'message' => 'Empty or bad id for question'], 400);
+        }
+
+        try {
+            $this->commandBus->handle(new BeginReplyHappeningQuestion($questionId, $user));
+        } catch(QuestionNotFoundException $e) {
+            return new JsonResponse(['status' => 'error', 'message' => $e->getMessage()], 404);
+        } catch(ReplyQuestionNotAllowedException $e) {
+            return new JsonResponse(['status' => 'error', 'message' => $e->getMessage()], 403);
+        } catch(HappeningException $e) {
+            return new JsonResponse(['status' => 'error', 'message' => $e->getMessage()], 400);
+        }
+
+        return new JsonResponse(
+            [
+                'status' => 'ok',
+            ]
+        );
+    }
+}